documentation changes, added git readme and walkthrough
